+++ Modules/Trackers.lua
@@ -113,8 +113,7 @@ function addon:GetTrackerCount(widget, trackerKey, overrideObjective)
     local count
 
     if trackerType == "ITEM" then
-        -- count = GetItemCount(trackerID, trackerInfo.includeBank)
-        count = trackerInfo.includeAllChars and self:GetDataStoreCount(trackerID, trackerInfo.includeBank) or GetItemCount(trackerID, trackerInfo.includeBank)
+        count = trackerInfo.includeAllChars and self:GetDataStoreItemCount(trackerID, trackerInfo.includeBank) or GetItemCount(trackerID, trackerInfo.includeBank)
     elseif trackerType == "CURRENCY" then
         count = GetCurrencyInfo(trackerID) and GetCurrencyInfo(trackerID).quantity
     end