+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bug fixes:
   - -- Fixed error when adding an objective with no trackers to a bar
   - -- Fixed errors preventing custom conditions from being set and used (may be reworked in the future)
+  - -- Added missing optional dependencies for includeAllChars: DataStore_Character, DataStore_Currencies
 
 ### 3.0-alpha9
 - Reimplemented quick add editbox